Use Xcode provided Clang and Binutils when compiling with iOS sdk
First, clang and libcxx seem to evolve in unisong, and particularly
mixing newer versions of libcxx with older version of clang causes
trouble.  Furthermore, Apple really does not expect you to mix and
match SDKs and compilers and doing otherwise just ends up banging your
head against the wall.

This commit instructs Nix to use the XCode provided tools when
cross-compiling to iOS.  The way this is done is very simple and
stupid, and my Nix knowledge is not deep enough to be sure whether
this is actually really bad.  This however has been surprisingly
effective in getting a lot of packages to compile.